Move platform admin ops ownership into smart core
