+++ addons/smart_core/controllers/__init__.py
@@ -1,3 +1,4 @@
 from . import intent_dispatcher
 from . import platform_menu_api
+from . import platform_ops_controller
 from ..app_config_engine.controllers import contract_api
